renamed core.configs to core.components
