+++ producer/spooling.go
@@ -22,7 +22,7 @@ import (
 	"time"
 
 	"github.com/trivago/gollum/core"
-	"github.com/trivago/gollum/core/configs"
+	"github.com/trivago/gollum/core/components"
 	"github.com/trivago/tgo"
 )
 
@@ -85,15 +85,15 @@ import (
 // default.
 type Spooling struct {
 	core.BufferedProducer `gollumdoc:"embed_type"`
-	outfile               map[core.MessageStreamID]*spoolFile
-	outfileGuard          *sync.RWMutex
-	rotation              configs.RotateConfig `gollumdoc:"embed_type"`
-	path                  string               `config:"Path" default:"/var/run/gollum/spooling"`
-	maxFileSize           int64                `config:"MaxFileSizeMB" default:"512" metric:"mb"`
-	batchMaxCount         int                  `config:"Batch/MaxCount" default:"100"`
-	bufferSizeByte        int                  `config:"BufferSizeByte" default:"8192"`
-	revertOnDrop          bool                 `config:"RevertStreamOnDrop"`
-	respoolDuration       time.Duration        `config:"RespoolDelaySec" default:"10" metric:"sec"`
+	outfile         map[core.MessageStreamID]*spoolFile
+	outfileGuard    *sync.RWMutex
+	rotation        components.RotateConfig `gollumdoc:"embed_type"`
+	path            string               `config:"Path" default:"/var/run/gollum/spooling"`
+	maxFileSize     int64                `config:"MaxFileSizeMB" default:"512" metric:"mb"`
+	batchMaxCount   int                  `config:"Batch/MaxCount" default:"100"`
+	bufferSizeByte  int                  `config:"BufferSizeByte" default:"8192"`
+	revertOnDrop    bool                 `config:"RevertStreamOnDrop"`
+	respoolDuration time.Duration        `config:"RespoolDelaySec" default:"10" metric:"sec"`
 	maxFileAge            time.Duration        `config:"MaxFileAgeMin" default:"1" metric:"min"`
 	batchTimeout          time.Duration        `config:"Batch/TimeoutSec" default:"5" metric:"sec"`
 	readDelay             time.Duration
@@ -136,7 +136,7 @@ func (prod *Spooling) Configure(conf core.PluginConfigReader) {
 	}
 
 	//TODO: check if rotation is still in use
-	prod.rotation = configs.NewRotateConfig()
+	prod.rotation = components.NewRotateConfig()
 	prod.rotation.Enabled = true
 	prod.rotation.Timeout = prod.maxFileAge
 	prod.rotation.SizeByte = prod.maxFileSize
